Move list guard/inner list to list.rs
